+++ src/app/utils/compressUtils.js
