+++ src/app/utils/hashUtils.js
